tests/wpt/web-platform-tests/css/css-tables/parsing/table-layout-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module: getComputedStyle().tableLayout</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-table-layout">
<meta name="assert" content="table-layout computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("table-layout", "auto");
test_computed_value("table-layout", "fixed");
</script>
</body>
</html>
